+++ CHANGELOG.md
@@ -8,6 +8,12 @@
   - Which takes from the end of an array instead of the beginning
 - Added new `dropLast` function
   - Returns a list containing all but the last n elements of the given list
+- Added new `pick` function
+  - Picks only the requested keys from a provided object
+
+### Improved
+
+- Updated `types` for `pathOr` so that it can be specified if needed (still defaults to `any`)
 
 ### Fixed